requirements_api.txt
# Django REST Framework API Dependencies

djangorestframework>=3.16.1,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.9.0,<5.0.0
django-filter>=25.1,<26.0
markdown>=3.9,<4.0

# API Documentation (Swagger/ReDoc)
drf-yasg[validation]>=1.21.11,<2.0.0
inflection>=0.5.1
ruamel.yaml>=0.17.0
uritemplate>=4.1.1
